+++ examples/direct-integration/src/App.js
@@ -1,6 +1,5 @@
 import { useEffect, useState } from 'react';
-import { On, Off, Heading, Button, Box, Text, Stack } from './components';
-import { ethers } from 'ethers';
+import { On, Off, Heading, Button, Box, Stack } from './components';
 import { loadConnectKit, SupportedProviders } from '@ledgerhq/connect-kit-loader';
 
 export const shortenAddress = (address) => {
@@ -15,59 +14,134 @@ export const shortenAddress = (address) => {
 
 export default function Home() {
   const [provider, setProvider] = useState();
-  const [library, setLibrary] = useState();
   const [account, setAccount] = useState();
   const [chainId, setChainId] = useState();
-  const [error, setError] = useState("");
+  const [message, setMessage] = useState("");
 
-  const connectWallet = async (chainId = 1) => {
+  const connectWallet = async () => {
+    console.log('> connectWallet');
+
+    resetState();
+
     try {
       const connectKit = await loadConnectKit();
       connectKit.enableDebugLogs();
       const checkSupportResult = connectKit.checkSupport({
-        chainId,
+        chainId: 137,
         providerType: SupportedProviders.Ethereum,
         rpc: {
-          1: `https://cloudflare-eth.com/`, // Mainnet
-          5: 'https://goerli.optimism.io',  // Goerli
-          137: "https://polygon-rpc.com/",  // Polygon
+          1: 'https://cloudflare-eth.com/',  // Mainnet
+          5: 'https://goerli.optimism.io/',  // Goerli
+          137: 'https://polygon-rpc.com/',   // Polygon
         }
       });
       console.log('checkSupportResult is', checkSupportResult);
 
-      const provider = await connectKit.getProvider();
-      setProvider(provider);
+      const connectKitProvider = await connectKit.getProvider();
+      setProvider(connectKitProvider);
 
-      const accounts = await provider.request({ method: 'eth_requestAccounts' });
-      if (accounts) setAccount(accounts[0]);
+      const requestAccountsResponse = await (requestAccounts(connectKitProvider));
+      if (requestAccountsResponse) setAccount(requestAccountsResponse[0]);
 
-      const library = new ethers.providers.Web3Provider(provider);
-      setLibrary(library);
-
-      const network = await library.getNetwork();
-      setChainId(network.chainId);
+      const chainIdResponse = await getChainId(connectKitProvider);
+      if (chainIdResponse) setChainId(chainIdResponse);
     } catch (error) {
-      setError(error);
+      console.error(error)
+      setMessage(error);
     }
   };
 
-  const disconnect = async () => {
+  const disconnectWallet = async () => {
+    console.log('> disconnectWallet');
+
+    if (provider.disconnect) {
+      console.log('> calling provider.disconnect()');
+      provider.disconnect();
+    }
+
+    // needs to be called here
+    // the Extension does not emit a disconnect event
+    resetState();
+  }
+
+  const resetState = async () => {
+    console.log('> resetState');
+
     setAccount();
     setChainId();
     setProvider();
-    setLibrary();
+    setMessage();
   };
 
+  const requestAccounts = async (provider) => {
+    try {
+      return await provider.request({ method: 'eth_requestAccounts' });
+    } catch (error) {
+      console.error(error)
+      setMessage(error);
+    }
+  }
+
+  const getChainId = async (provider) => {
+    try {
+      return await provider.request({ method: 'eth_chainId' });
+    } catch (error) {
+      console.error(error)
+      setMessage(error);
+    }
+  }
+
+  const isRequestedChainId = (requestedChainId) => {
+    console.log(`comparing ${requestedChainId} to ${chainId}`)
+    return chainId === requestedChainId || chainId === `0x${requestedChainId.toString(16)}`;
+  }
+
+  const switchChains = async (chainId) => {
+    try {
+      await provider.request({
+        method: 'wallet_switchEthereumChain',
+        params: [{ chainId: `0x${chainId.toString(16)}` }],
+      });
+      setMessage();
+    } catch (error) {
+      console.error(error)
+      setMessage(error);
+    }
+  }
+
   useEffect(() => {
+    console.log('> useEffect');
+
     if (provider?.on) {
-      const handleDisconnect = (error) => {
-        disconnect();
+      const handleDisconnect = (props) => {
+        console.log('> handleDisconnect', props);
+
+        // needs to be called here to handle disconnect from Ledger Live
+        resetState();
       };
 
+      const handleChainChanded = async (chainId) => {
+        console.log('> handleChainChanded', chainId);
+        setChainId(chainId);
+      }
+
+      const handleAccountsChanged = async (accounts) => {
+        console.log('> handleAccountsChanged', accounts);
+        setAccount(accounts[0]);
+      }
+
+      provider.on('chainChanged', handleChainChanded);
+      provider.on('accountsChanged', handleAccountsChanged);
       provider.on("disconnect", handleDisconnect);
 
       return () => {
+        // handle removing event listeners here,
+        // when disconnecting from the Extension no disconnect event is emited
         if (provider.removeListener) {
+          console.log('> removing event listeners');
+
+          provider.removeListener('chainChanged', handleChainChanded);
+          provider.removeListener('accountsChanged', handleAccountsChanged);
           provider.removeListener("disconnect", handleDisconnect);
         }
       };
@@ -86,20 +160,34 @@ export default function Home() {
 
         {account && (
           <>
-            <Box>{`Network Id: ${chainId ? chainId : "none"}`}</Box>
+            <Box>{`Chain Id: ${chainId ? chainId : "none"}`}</Box>
             <Box>{`Account: ${shortenAddress(account)}`}</Box>
           </>
         )}
 
-        <Box>{error ? error.message : null}</Box>
+        <Box>{message ? message.message : null}</Box>
 
-        <Box>
-          {!account ? (
-            <Button bg='primary' onClick={() => connectWallet(1)}>Connect Wallet</Button>
-          ) : (
-            <Button onClick={disconnect}>Disconnect</Button>
-          )}
-        </Box>
+        {!account ? (
+          <Box>
+            <Button bg='primary' onClick={connectWallet}>Connect Wallet</Button>
+          </Box>
+        ) : (
+          <>
+            <Box>
+              <Button onClick={disconnectWallet}>Disconnect</Button>
+            </Box>
+
+            <Box>
+              <Button disabled={isRequestedChainId(137)} onClick={() => switchChains(137)}>Switch to Polygon</Button>
+            </Box>
+            <Box>
+              <Button disabled={isRequestedChainId(5)} onClick={() => switchChains(5)}>Switch to Göerli</Button>
+            </Box>
+            <Box>
+              <Button disabled={isRequestedChainId(1)} onClick={() => switchChains(1)}>Switch to Mainnet</Button>
+            </Box>
+            </>
+        )}
       </Stack>
     </>
   );