+++ packages/connect-kit/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Remove the "Need a Ledger?" section from the modals.
+- Replace Connect with Ledger Extension.
+- Center the modal.
 
 ## 1.0.11 - 2023-02-24
 ### Changed